Revert "Revert 4482c19..3136142"
This reverts commit e48a5596c99f967b1dfaec6b679b95665776b8f9.

+++ UOless/Balance.swift
@@ -11,10 +11,10 @@ import Foundation
 class Balance {
 	var contact: Contact
 	var currency: Currency
-	var balance: Float //Positive = get, negative = have to pay
+	var balance: Double //Positive = get, negative = have to pay
 	var unprocessed: Bool //if true, there are some transactions not yet canceled out, so there will probably be room for improvement, which can be seen when refreshing (canceling happens every 10 min)
 	
-	init(contact: Contact, currency: Currency, balance: Float, unprocessed: Bool) {
+	init(contact: Contact, currency: Currency, balance: Double, unprocessed: Bool) {
 		self.contact = contact
 		self.currency = currency
 		self.balance = balance
